// web/src/app/core/models/enums/NatureEnum.ts
export enum NatureEnum {
  BASHFUL= 'Bashful',
  DOCILE = 'Docile',
  HARDY = 'Hardy',
  QUIRKY = 'Quirky',

  ADAMANT = 'Adamant',
  BRAVE = 'Brave',
  LONELY = 'Lonely',
  NAUGHTY = 'Naughty',
  BOLD = 'Bold',
  IMPISH = 'Impish',
  LAX = 'Lax',
  RELAXED = 'Relaxed',
  MODEST = 'Modest',
  MILD = 'Mild',
  RASH = 'Rash',
  QUIET = 'Quiet',
  CALM = 'Calm',
  GENTLE = 'Gentle',
  CAREFUL = 'Careful',
  SASSY = 'Sassy',
  TIMID = 'Timid',
  HASTY = 'Hasty',
  JOLLY = 'Jolly',
  NAIVE = 'Naive'
}